+++ src/spellbot/cogs/events_cog.py
@@ -8,7 +8,7 @@ from discord.ext import commands
 
 from spellbot import SpellBot
 from spellbot.actions import LookingForGameAction
-from spellbot.enums import GAME_FORMAT_ORDER, GAME_SERVICE_ORDER
+from spellbot.enums import GAME_BRACKET_ORDER, GAME_FORMAT_ORDER, GAME_SERVICE_ORDER
 from spellbot.metrics import add_span_context
 from spellbot.operations import safe_defer_interaction
 from spellbot.settings import settings
@@ -29,6 +29,10 @@ class EventsCog(commands.Cog):
     @app_commands.choices(
         format=[Choice(name=str(format), value=format.value) for format in GAME_FORMAT_ORDER],
     )
+    @app_commands.describe(format="What commander bracket do you want to play?")
+    @app_commands.choices(
+        bracket=[Choice(name=str(bracket), value=bracket.value) for bracket in GAME_BRACKET_ORDER],
+    )
     @app_commands.describe(service="What service do you want to use to play this game?")
     @app_commands.choices(
         service=[Choice(name=str(service), value=service.value) for service in GAME_SERVICE_ORDER],
@@ -39,13 +43,14 @@ class EventsCog(commands.Cog):
         interaction: discord.Interaction,
         players: str,
         format: int | None = None,
+        bracket: int | None = None,
         service: int | None = None,
     ) -> None:
         assert interaction.channel_id is not None
         add_span_context(interaction)
         await safe_defer_interaction(interaction)
         async with LookingForGameAction.create(self.bot, interaction) as action:
-            await action.create_game(players, format, service)
+            await action.create_game(players, format, bracket, service)
 
 
 async def setup(bot: SpellBot) -> None:  # pragma: no cover