+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds support for the Commander Bracket system.
+
 ## [v14.5.1](https://github.com/lexicalunit/spellbot/releases/tag/v14.5.1) - 2025-02-02
 
 ### Fixed